Clean up service package
