+++ service/doc.go
@@ -78,8 +78,8 @@
 //       "kafka",
 //       kafka.Module("kakfa_topic1", []string{"worker"}),
 //     ).WithModule(
-//       "keyvalue",
-//       yarpc.New(rpc.CreateThriftServiceFunc(NewYarpcThriftHandler)),
+//       "yarpc",
+//       yarpc.New(yarpc.CreateThriftServiceFunc(NewYarpcThriftHandler)),
 //       service.WithModuleRole("service"),
 //     ).Build()
 //
